config: Remove Appengine config
Remove Appengine config, because 4$ a day for a toy project is
unjustifiable. This is only the cost of Appengine flex, with a
permanently running instance, which is the required setup for websockets
and the essence of what we set out to achieve with foxtrot (operational
transforms and instant editing collaboration).

Julia got a Pi in the meantime and is getting over her fears of
self-hosting. Cam never had those.