pkg: Add room and message db layer
Add room and message db layer to create and get a room by id and create
a new message  or query messages for a given room with an optional
message count `limit` and an optional `beforeID` index for pagination in
the UI.

It turns out that sqlite doesn't enable foreign key constraints by
default. To enable foreign key constraints with the given sqlite3 driver
add use `":memory:?_foreign_keys=true"` for an in-memory db as used in a
MessageQuery test.