sql: Embed SQL files using Go 1.16 embedding
Use Go 1.16 file embedding to embed the SQL files, rather than
generating them as Go strings. This fixes the problem where the SQL file
cannot contain a backquote otherwise it breaks the Go string. This could
be fixed with a fancier generator, but instead now that the Go 1.16 beta
is out with embedding, switch to that instead.