lint: Change lint version to 1.33.2
Change the lint version to 1.33.2 and when running from CI, build that
version so it is used locally instead of running it from docker
(`lint-with-docker`). This makes golangci-lint use the version of Go
that is setup on the build agent and not the version bundled with the
docker image. It is not possible to use the golangci-lint github action
as it also installs a version of Go, overriding whatever you have setup
on the builder.

This also seems to have sped up the build - it is 10s faster by the
looks of it.