deploy: Add k8s deployment
Add k8s jsonnet deployment, parameterised on hostname and docker tag.

Usage:

    kubecfg update -A hostname=foxtrot.example.com -A docker_tag=v0.0.1 main.jsonnet

Note: If hostname is set an ingress resource is created for which it
assumes that Let's encrypt is set up on the cluster and that there is a
Traefik entry point called https.

If no docker tag is provided, the default `latest` is used.

Example deployment with k3s setup [1]:

    export KUBECONFIG=/opt/k3s/etc/kubeconfig.yaml
    kubecfg update -A hostname=foxtrot.example.com main.jsonnet

Use TLAs to configure the manifests instead of extVars, so we can have
default values that can be optionally overridden.

[1]: https://github.com/foxygoat/k3sinst