github: Rework CI/CD workflow
Rework CI/CD GitHub action workflow. Merge ci.yaml and release.yaml
files so that releasing a new image will only happen if the CI run
(make, make docker-test) is success. Jobs can only depend on each other
within the same workflow and yaml file, using the `needs:` field.

As a further step, separate releasing and deploying, as we wish to
deploy for PRs as well as for merges to master in a follow up commit, so
that we can easily review and preview UI changes, *before* they get
merged to master.

Only new commits on master get "release", i.e.. git-tagged, a new github
Release page, and `get` from pkg.go.dev.

However, new commits on master as well as PRs with label `deploy` have a
docker image built and pushed to dockerhub. In a follow up step we want
to trigger a k8s deployment with the new docker_tag.