make: Replace $(REF) with $(COMMIT_SHA)
Replace $(REF) with $(COMMIT_SHA). REF is used in raw github URLs to
refer to the exact version of "remote" jsonnet files to be used by JCDC.
Initially we intended to use the branch name (hence the name REF) but it
turned out to be easier to use the commit sha. However, we never updated
the corresponding environment variable so we ended up using duplicating
the COMMIT_SHA passed in - simplify by renaming REF to COMMIT_SHA.