deploy: Annotate deployment with commit sha
Annotate deployment template spec with commit sha so that a pod is
re-deployed if nothing in deployment has changed, for instance if a PR
got update the manifest will stay the same with same label PRNN, however
we want to trigger re-deploy. Annotating the deployment template spec
with the commitSha will ensure a redeployment on any code change or even
empty force push.