ui: Purge unused svelte boilerplate
Purge unused svelte boilerplate from frontend/:

	- README.md
	- .gitignore
	- global.css
	- favicon.png

The README.md and .gitignore at the repo root now cover frontend/ as
well. In a follow up commit we will use TailwindCSS for styles so there
is no need for a global style sheet.

Update index.html to use Foxtrot as the app Title and render a fox emoji
as favicon.