deploy: (Un)deploy only dev-specific resources on PRs
Exclude any resources in the manifest list that are common to all
deployments when doing a dev deployment.

The PR (branch) deployments were also deploying the resources that were
not custom per branch (the namespace and the sealed secret). This meant
when undeploying a branch deployment, it would also remove those shared
resources. So we added a hack for dev undeployment.

We can now remove this hack as the un/deployment isolates just the
resources specific to that deployment. This will make it easier to clean
up when a PR is closed.