CI: Add undeploy workflow to undeploy on jul.run
Add a workflow triggering off a pull request being closed that undeploys
the PR deployment on jul.run. This means if a PR is closed or merged and
it was deployed, it will be cleaned up.

Still to do is delete the image/tag from docker hub. This would probably
best be done with a custom action. We could open-code some API calls but
the secrets management is a little dodge, and probably safer just making
an action.